Quicksave the markdown
I made a lot of progress, and we are in the middle of some blackouts so I want to save this. Still not done yet, and it doesn't have the same mindset I have now with where this language is going. Arguably this is more important than the code, because with this I can write the code probably twice as fast and keep true to my goals.