Wrote my thoughts out before I blackout
TODO: Everything

And edit the makefile so it can compile both c and cpp, and without me
thinking about it

Also research if c object files (Why would they call them that if there
is no objects) can be compiled with c files. That way I might just be
able to have a main.cpp or something. Damn I'm tired