Just testing out my buffer. It works, kinda
It only works if you pipe in the file, because it needs that EOF to get
out of the first character loop.

In cpp this wouldn't be a problem because they got things like getline.

I could still try that character by character stack thing, which is
really why I'm pushing this commit.