Basically doing my Leifix inside of c++
There's some cool things happening. Turns out the easiest way to make a
variable is to define it as a function that returns whatever data it
holds. I kind of already knew that, but it's cool to see it work.

Another thing is that I got a lot of this right the first time, the only
thing that was off was I thought the variables were held on a stack, and
I had no idea about the operator queue.